Libi udimi

+++ src/main/java/ch/uzh/ifi/seal/soprafs19/service/ActionService.java
@@ -42,7 +42,13 @@ public class ActionService {
         action.perfromAction(this.gameService);
         long gameId = action.TheGameId();
         Game game = this.gameService.gameByID(gameId);
+        if( game == null){
+            return true;
+        }
         game.addAction(action);
+        if(!action.needToDelete()){
+            return true;
+        }
         if( game.retriveActions1() != null) {
             for (long actionId : game.retriveActions1()) {
                 this.actionRepository.deleteById(actionId);
